File Restructure
This PR restructures the existing program. Each class has been split into it own Java file
