+++ src/main/java/Main.java
@@ -1,0 +1,8 @@
+package src.main.java;
+
+public class Main {
+    public static void main(String[] args) {
+        ZombieWar apocalypse = new ZombieWar(2, 1);
+        apocalypse.simulate();
+    }
+}
