+++ src/main/java/characters/humans/Civilian.java
@@ -1,0 +1,7 @@
+package src.main.java.characters.humans;
+
+public class Civilian extends Survivor {
+    public Civilian() {
+        super(50,5);
+    }
+}
