+++ src/main/java/characters/humans/Scientist.java
@@ -1,0 +1,7 @@
+package src.main.java.characters.humans;
+
+public class Scientist extends Survivor {
+    public Scientist() {
+        super(20, 2);
+    }
+}
