+++ src/main/java/characters/humans/Soldier.java
@@ -1,0 +1,7 @@
+package src.main.java.characters.humans;
+
+public class Soldier extends Survivor {
+    public Soldier() {
+        super(100, 10);
+    }
+}
