+++ src/main/java/characters/humans/Survivor.java
@@ -1,0 +1,9 @@
+package src.main.java.characters.humans;
+
+import src.main.java.characters.Player;
+
+public class Survivor extends Player {
+    public Survivor(int health, int power) {
+        super(health, power);
+    }
+}
