+++ src/main/java/characters/zombies/CommonInfected.java
@@ -1,0 +1,7 @@
+package src.main.java.characters.zombies;
+
+public class CommonInfected extends Zombie {
+    public CommonInfected() {
+        super(30,5);
+    }
+}
