+++ src/main/java/characters/zombies/Tank.java
@@ -1,0 +1,7 @@
+package src.main.java.characters.zombies;
+
+public class Tank extends Zombie {
+    public Tank() {
+        super(150, 20);
+    }
+}
