+++ src/main/java/characters/zombies/Zombie.java
@@ -1,0 +1,9 @@
+package src.main.java.characters.zombies;
+
+import src.main.java.characters.Player;
+
+public class Zombie extends Player {
+    public Zombie(int health, int power) {
+        super(health, power);
+    }
+}
